break event observer into it's own file
